+++ frontend/src/pages/FormPages/EmailSent.jsx
