+++ frontend/src/pages/FormPages/Register.jsx
@@ -1,15 +1,15 @@
-import Form from "../components/Form";
+import Form from "../../components/Form";
 
 function Register() {
   return (
-    <>
+    <div>
       <img src="/logo.svg" alt="AI Logo" className="logo" />
       <Form
         route="/api/user/register/"
         method="register"
         title="Start Today!"
       />
-    </>
+    </div>
   );
 }
